Moved auth0ProviderWithRedirect.tsx
